refactor(workitems): extract fields into nx_lib/workitems
Move the field-config builder (_build_field_config/_LABEL_LANGS) and the
doc-value search DSL (DOCFIELD_OPS, _docfield_op/_docfield_comb,
_docfield_predicate, _docfield_pairs_normalized, _docfield_ids_cache_key)
into nx_lib/workitems/fields.py -- pure Python, no Flask/session/request.

views/workitems.py now imports these from the new module; _docfield_op/
_docfield_comb keep thin module-qualified forwarders so the existing
_docfield_op spy test (adapted to patch nx_lib.workitems.fields directly)
keeps working once the pair-fold loop itself moves to
nx_lib.workitems.query in a later commit.
